Revert "Simplify"
This reverts commit 212becf4f5a0a9d430ea8a5a82f9c3f6b3c36746.
