--- test-nextjs/app/signin-test/page.tsx
@@ -1,58 +1,0 @@
-"use client";
-
-import { GitHubLogo } from "@/components/GitHubLogo";
-import { Button } from "@/components/ui/button";
-import { Input } from "@/components/ui/input";
-import { useAuthActions } from "@convex-dev/auth/react";
-import Link from "next/link";
-import { useRouter } from "next/navigation";
-
-// Only used in automated e2e tests
-export default function SignTest() {
-  return (
-    <div className="flex min-h-screen w-full">
-      <main className="mx-auto my-auto flex-col flex">
-        <h2 className="font-semibold text-2xl tracking-tight mb-4">
-          Sign in or create an account
-        </h2>
-        <SignInWithSecret />
-
-        <Button variant="link" className="text-muted-foreground" asChild>
-          <Link href="/">Cancel</Link>
-        </Button>
-      </main>
-    </div>
-  );
-}
-
-function SignInWithSecret() {
-  const { signIn } = useAuthActions();
-  const router = useRouter();
-  return (
-    <form
-      className="flex flex-col gap-2 mt-8"
-      onSubmit={(event) => {
-        event.preventDefault();
-        const formData = new FormData(event.currentTarget);
-        signIn("secret", formData)
-          .then(() => {
-            router.push("/product");
-          })
-          .catch(() => {
-            window.alert("Invalid secret");
-          });
-      }}
-    >
-      Sign in with a secret
-      <Input
-        aria-label="Secret"
-        type="text"
-        name="secret"
-        placeholder="secret value"
-      />
-      <Button className="flex-1" variant="outline" type="submit">
-        Sign in with secret
-      </Button>
-    </form>
-  );
-}
