+++ test-nextjs/app/signin/page.tsx
@@ -1,9 +1,11 @@
 "use client";
 
-import { useAuthActions } from "@convex-dev/auth/react";
 import { GitHubLogo } from "@/components/GitHubLogo";
 import { Button } from "@/components/ui/button";
+import { Input } from "@/components/ui/input";
+import { useAuthActions } from "@convex-dev/auth/react";
 import Link from "next/link";
+import { useRouter } from "next/navigation";
 
 export default function Sign() {
   return (
@@ -13,6 +15,7 @@ export default function Sign() {
           Sign in or create an account
         </h2>
         <SignInWithGitHub />
+        {process.env.NEXT_PUBLIC_E2E_TEST && <SignInWithSecret />}
 
         <Button variant="link" className="text-muted-foreground" asChild>
           <Link href="/">Cancel</Link>
@@ -33,5 +36,41 @@ function SignInWithGitHub() {
     >
       <GitHubLogo className="mr-2 h-4 w-4" /> GitHub
     </Button>
+  );
+}
+
+// Only used in automated e2e tests
+function SignInWithSecret() {
+  const { signIn } = useAuthActions();
+  const router = useRouter();
+  return (
+    <form
+      className="flex flex-col gap-2 mt-8"
+      onSubmit={(event) => {
+        console.log("signin with secret");
+        event.preventDefault();
+        const formData = new FormData(event.currentTarget);
+        signIn("secret", formData)
+          .then(() => {
+            console.log("signin with secret success");
+            router.push("/product");
+          })
+          .catch(() => {
+            console.log("signin with secret failure");
+            window.alert("Invalid secret");
+          });
+      }}
+    >
+      Test only: Sign in with a secret
+      <Input
+        aria-label="Secret"
+        type="text"
+        name="secret"
+        placeholder="secret value"
+      />
+      <Button className="flex-1" variant="outline" type="submit">
+        Sign in with secret
+      </Button>
+    </form>
   );
 }
