+++ test-nextjs/e2e-tests/README.md
@@ -4,14 +4,15 @@
 
    `npx convex run tests:init`
 
-2. Set up the secret on Convex backend:
+1. Set up the secret on Convex backend:
 
    `npx convex env set AUTH_E2E_TEST_SECRET <something>`
 
-3. Specify the secret for playwright:
+1. Turn on the secret sign-in UI and specify the secret for playwright:
 
    Add the following to `.env.local`:
 
    ```
+   NEXT_PUBLIC_E2E_TEST=true
    AUTH_E2E_TEST_SECRET=<something>
    ```
