+++ test-nextjs/e2e-tests/signin.spec.ts
@@ -1,7 +1,9 @@
 import { test, expect } from "@playwright/test";
 
 test("signin fails correctly", async ({ page }) => {
-  await page.goto("/signin-test");
+  await page.goto("/");
+
+  await page.getByRole("link").getByText("Get Started").first().click();
 
   await page
     .getByLabel("Secret")
