+++ test-nextjs/e2e-tests/signout.spec.ts
@@ -1,7 +1,9 @@
 import { test, expect } from "@playwright/test";
 
 test("signout works", async ({ page }) => {
-  await page.goto("/signin-test");
+  await page.goto("/signin");
+
+  // await page.getByRole("link").getByText("Get Started").first().click();
 
   await page.getByLabel("Secret").fill(process.env.AUTH_E2E_TEST_SECRET!);
